pyproject.toml
[project]
name = "autonomous-agent"
version = "0.1.0"
description = "Autonomous agent with tools and skills"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "tenacity>=9.0.0",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.1",
    "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "httpx", "ruff"]
openai = ["openai>=1.55.0"]

[project.scripts]
agent = "agent.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
